+++ sql/types/geometry.go
@@ -546,7 +546,7 @@ func (t GeometryType) MatchSRID(v interface{}) error {
 }
 
 func ValidateSRID(srid int, funcName string) error {
-	if srid < 0 || srid > math.MaxUint32 {
+	if srid < 0 || int64(srid) > int64(math.MaxUint32) {
 		return sql.ErrInvalidSRID.New(funcName)
 	}
 	if _, ok := SupportedSRIDs[uint32(srid)]; !ok {
